[AWIBOF-7346] merge template specializations into one header and remove zero-padding for proto serialization
