--- src/allocator/context_manager/context/context_section_base.h
@@ -1,78 +1,0 @@
-/*
-*   BSD LICENSE
-*   Copyright (c) 2022 Samsung Electronics Corporation
-*   All rights reserved.
-*
-*   Redistribution and use in source and binary forms, with or without
-*   modification, are permitted provided that the following conditions
-*   are met:
-*
-*     * Redistributions of source code must retain the above copyright
-*       notice, this list of conditions and the following disclaimer.
-*     * Redistributions in binary form must reproduce the above copyright
-*       notice, this list of conditions and the following disclaimer in
-*       the documentation and/or other materials provided with the
-*       distribution.
-*     * Neither the name of Samsung Electronics Corporation nor the names of
-*       its contributors may be used to endorse or promote products derived
-*       from this software without specific prior written permission.
-*
-*   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-*   "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-*   LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-*   A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-*   OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-*   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-*   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-*   DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-*   THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-*   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-*   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
-#pragma once
-namespace pos
-{
-template<typename T>
-class ContextSectionBase
-{
-public:
-    T data;
-    
-    void InitAddressInfo(uint64_t offset, uint64_t size)
-    {
-        info.offset = offset;
-        info.size = size;
-    }
-
-    /**
-     * "buf" points to the start address of the continguous memory region passed in to "SegmentCtx::BeforeFlush()"
-     */
-    virtual void CopyTo(char* buf)
-    {
-        char* destBuf = buf + this->info.offset;
-        data->ToBytes(destBuf);
-    }
-
-    /**
-     * "buf" points to the start address of the continguous memory region passed in to "SegmentCtx::AfterLoad()"
-     */
-    virtual void CopyFrom(char* buf)
-    {
-        char* srcBuf = buf + this->info.offset;
-        data->FromBytes(srcBuf);
-    }
-
-    virtual uint64_t GetSectionSize(void)
-    {
-        return data->SerializedSize();
-    }
-
-    ContextSectionAddr GetSectionInfo(void)
-    {
-        return info;
-    }
-protected:
-    ContextSectionAddr info;
-};
-}
