+++ src/allocator/context_manager/segment_ctx/segment_info.cpp
@@ -309,12 +309,6 @@ SegmentInfoData::ToBytes(char* destBuf)
             return false;
         }
 
-        // fill zeros up to ONSSD_SIZE bytes-position
-        for(unsigned int bytePos = effectiveSize; bytePos < ONSSD_SIZE; bytePos ++)
-        {
-            destBuf[bytePos] = 0;
-        }
-
         return true;
     }
     else
